refactor: file controller, service 리팩토링
